pyproject.toml
[tool.poetry]
name = "alert-hub-backend"
version = "0.1.0"
description = ""
authors = ["IFRC"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
celery = "^5.3"
django-celery-beat = "*"
django-cors-headers = "^4.2.0"
django-environ = "*"
django-extensions = "^3.2.3"
django-redis = "^5.3.0"
django-storages = "^1.13.2"
iso639-lang = "^2.1.0"
graphene-django = "*"
psycopg2-binary = "^2.9.6"
shapely = "^2.0.1"
pytz = "*"
colorlog = "*"
requests = "*"
strawberry-graphql = "^0.220.0"
strawberry-django-plus = "*"
sentry-sdk = "*"
djangorestframework = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
ipython = "*"
pytest-profiling = "*"
pytest-ordering = "*"
pytest-django = "*"
django-stubs = "==4.2.6"  # NOTE: 4.2.7 is messing with TextChoices/IntergerChoices
celery-types = "*"

[tool.black]
line-length = 125
skip-string-normalization = true
# NOTE: Update in .pre-commit-config.yaml as well
extend-exclude = "(__pycache__|.*snap_test_.*\\.py|.+\\/.+\\/migrations\\/.*)"

[tool.isort]
profile = "black"

[tool.pyright]
exclude = [
    "**/__pycache__",
    "**/snap_test_*.py",
    ".venv/",
    "legacy/",
    "**/migrations/*.py",
]
reportMissingImports = true
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "main.settings"
log_cli = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
